move functions to file
